Revert "UNSTABLE - Added a few fixes & improvements"
This reverts commit 90b113ca15fee886ad40aa35c616eb0055ee251c.

+++ src/core_elements.h
@@ -90,8 +90,8 @@ struct Point {
 };
 
 struct IntPoint {
-	short x;
-	short y;
+	int x;
+	int y;
 };
 
 struct Scale {
